Refix gcc pickiness fix to conform to haranp-dislikes-C-style-casts-pickiness :)
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2119 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/spells3.cc
@@ -766,7 +766,7 @@ bool entomb(int powc)
                 continue;
 
             bool proceed = false;
-            for (int i = 0; i < (int) ARRAYSIZE(safe_to_overwrite); ++i)
+            for (unsigned int i = 0; i < ARRAYSIZE(safe_to_overwrite); ++i)
             {
                 if (grd[srx][sry] == safe_to_overwrite[i])
                 {
